Add update form functinoality

+++ app/src/main/java/com/project/symptoms/db/controller/SelectedCategoryOptionController.java
@@ -50,7 +50,7 @@ public class SelectedCategoryOptionController {
         return result;
     }
 
-    public List<SelectedCategoryOptionModel> getAllBySymptom(int symptomId){
+    public List<SelectedCategoryOptionModel> getAllBySymptom(long symptomId){
         List<SelectedCategoryOptionModel> result = new ArrayList<>();
         try{
             result = selectedCategoryOptionDao.listAllBySymptom(symptomId);
